Resolved log argument type conflict in SipDialogMonitorTest
git-svn-id: https://sipxsvn.sipez.com/rep/sipX/main@12539 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXtackLib/src/test/net/SipDialogMonitorTest.cpp
@@ -1,7 +1,6 @@
 // 
 // 
-// Copyright (C) 2006-2010 SIPez LLC. 
-// Licensed to SIPfoundry under a Contributor Agreement. 
+// Copyright (C) 2006-2017 SIPez LLC.  All rights reserved.
 //
 // Copyright (C) 2006 SIPfoundry Inc.
 // Licensed by SIPfoundry under the LGPL license.
@@ -106,7 +105,7 @@ public:
                  "Content-Length: %d\r\n"
                  "\r\n",
                  seqNo++,
-                 strlen(body));
+                 (int)strlen(body));
          message.append(header);
          message.append(body);
